# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PassportVerificationStatus = typeof PassportVerificationStatus[keyof typeof PassportVerificationStatus];


export const PassportVerificationStatus = {
  pending: 'pending',
  verified: 'verified',
} as const;

export interface Passport {
  passportId: string;
  product: string;
  brand: string;
  model: string;
  serialNumber: string;
  category: string;
  documentType: string;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  purchasePrice?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  warranty?: string | null;
  /** @nullable */
  seller?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  invoiceNumber?: string | null;
  /** @nullable */
  sourceDocument?: string | null;
  /** @nullable */
  physicalProductImage?: string | null;
  /** @nullable */
  physicalScanDate?: string | null;
  /** @nullable */
  matchConfidence?: number | null;
  verificationStatus: PassportVerificationStatus;
  createdAt: string;
  updatedAt: string;
}

export interface PassportInput {
  product: string;
  brand: string;
  model: string;
  serialNumber: string;
  category: string;
  documentType: string;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  purchasePrice?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  warranty?: string | null;
  /** @nullable */
  seller?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  invoiceNumber?: string | null;
  /** @nullable */
  sourceDocument?: string | null;
}

export interface PassportUpdate {
  product?: string;
  brand?: string;
  model?: string;
  serialNumber?: string;
  category?: string;
  documentType?: string;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  purchasePrice?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  warranty?: string | null;
  /** @nullable */
  seller?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  invoiceNumber?: string | null;
}

export interface DocumentAnalysisInput {
  fileName: string;
  fileType: string;
  /** @nullable */
  content?: string | null;
}

export interface DetectedProduct {
  product: string;
  brand: string;
  model: string;
  serialNumber: string;
  category: string;
  selected: boolean;
  evidence: string;
}

export type DocumentAnalysisExtractedFields = {[key: string]: string | number | null};

export interface DocumentAnalysis {
  documentType: string;
  products: DetectedProduct[];
  extractedFields?: DocumentAnalysisExtractedFields;
}

export interface ProductIdentificationInput {
  image: string;
}

export interface ProductIdentification {
  detectedProduct: string;
  category: string;
  brand: string;
  model: string;
  serialNumber: string;
  confidence: number;
  boundingBox?: number[];
  visualFeatures?: string[];
}

export interface ProductMatchInput {
  detectedProduct: ProductIdentification;
}

export type ProductMatchConfidence = typeof ProductMatchConfidence[keyof typeof ProductMatchConfidence];


export const ProductMatchConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface ProductMatch {
  passportId: string;
  matchScore: number;
  confidence: ProductMatchConfidence;
  reason: string[];
  passport: Passport;
}

export interface ProductMatchResponse {
  matches: ProductMatch[];
}

export interface ProductLinkInput {
  image: string;
  confidence: number;
  scanDate?: string;
}

export interface Activity {
  id: string;
  type: string;
  title: string;
  description: string;
  timestamp: string;
  /** @nullable */
  passportId?: string | null;
}

export interface DashboardSummary {
  totalPassports: number;
  productsScanned: number;
  documentsProcessed: number;
  successfullyLinked: number;
}

export type ServiceStatusStatus = typeof ServiceStatusStatus[keyof typeof ServiceStatusStatus];


export const ServiceStatusStatus = {
  connected: 'connected',
  degraded: 'degraded',
  unavailable: 'unavailable',
} as const;

export interface ServiceStatus {
  name: string;
  status: ServiceStatusStatus;
  detail: string;
}

export interface SystemStatus {
  backend: ServiceStatus;
  services: ServiceStatus[];
}

export type ListPassportsParams = {
search?: string;
category?: string;
status?: string;
};