# packages/api-zod/src/generated/types/detectedProduct.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DetectedProduct {
  product: string;
  brand: string;
  model: string;
  serialNumber: string;
  category: string;
  selected: boolean;
  evidence: string;
}