# packages/api-zod/src/generated/types/documentAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DetectedProduct } from './detectedProduct';
import type { DocumentAnalysisExtractedFields } from './documentAnalysisExtractedFields';

export interface DocumentAnalysis {
  documentType: string;
  products: DetectedProduct[];
  extractedFields?: DocumentAnalysisExtractedFields;
}