# packages/api-zod/src/generated/types/passportUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PassportUpdate {
  product?: string;
  brand?: string;
  model?: string;
  serialNumber?: string;
  category?: string;
  documentType?: string;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  purchasePrice?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  warranty?: string | null;
  /** @nullable */
  seller?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  invoiceNumber?: string | null;
}